requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain
pydantic
langchain-core
